ManoirOublié/manoir-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
eventlet==0.33.3
python-dotenv==1.0.1
google-cloud-firestore==2.16.0
google-auth==2.33.0
pyjwt[crypto]==2.9.0